third_party/blink/web_tests/external/wpt/editing/run/insertorderedlist_1001-last-expected.txt

This is a testharness.js-based test.
Found 168 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>[bar]<li>baz</ul><li>quz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>[bar]<li>baz</ul><li>quz</ul>" queryCommandState("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>[bar]<li>baz</ul><li>quz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>bar<li>[baz]</ul><li>quz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul><ol><li>baz</li></ol><li>quz</li></ul>" but got "<ul><li>foo<ul><li>bar</li></ul><ol><li>baz</li></ol></li><li>quz</li></ul>"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>bar<li>[baz]</ul><li>quz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>bar<li>[baz]</ul><li>quz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo</li><ul><li>[bar]<li>baz</ul><li>quz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo</li><ul><li>[bar]<li>baz</ul><li>quz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo</li><ul><li>bar<li>[baz]</ul><li>quz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo</li><ul><li>bar<li>[baz]</ul><li>quz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul><li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><ul><ul><li>bar</li></ul><li>baz</li></ul>" but got "<ol><li>foo</li></ol><ul><li>bar</li></ul><ul><li>baz</li></ul>"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul><li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul><li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo]</li><ul><li>bar</ul><li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo]</li><ul><li>bar</ul><li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<li>[bar]<ul><li>baz</ul><li>quz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><ol><li>bar</li></ol><ul><ul><li>baz</li></ul><li>quz</li></ul>" but got "<ul><li>foo</li></ul><ol><li>bar</li></ol><ul><li>baz</li></ul><ul><li>quz</li></ul>"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<li>[bar]<ul><li>baz</ul><li>quz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<li>[bar]<ul><li>baz</ul><li>quz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<li>[bar]</li><ul><li>baz</ul><li>quz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<li>[bar]</li><ul><li>baz</ul><li>quz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>bar<li>baz</ul><li>[quz]</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>bar<li>baz</ul><li>[quz]</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo</li><ul><li>bar<li>baz</ul><li>[quz]</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo</li><ul><li>bar<li>baz</ul><li>[quz]</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<li>[bar<li>baz]</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<li>[bar<li>baz]</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><ul><li>baz</li></ul>" but got "<ol><li>foo</li><li>bar</li></ol><ul><li>baz</li></ul>"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul><li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li></ul><ol><ol><li>bar</li></ol><li>baz</li></ol>" but got "<ul><li>foo<ol><li>bar</li></ol></li></ul><ol><li>baz</li></ol>"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" queryCommandState("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>b[ar</ul><li>b]az</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo<ul><li>bar</ul><li>baz]</ul><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol><li>baz</li></ol><p>extra</p>" but got "<ol><li>foo<ol><li>bar</li></ol></li><li>baz</li></ol><p>extra</p>"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo<ul><li>bar</ul><li>baz]</ul><p>extra" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo<ul><li>bar</ul><li>baz]</ul><p>extra" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><ul><ul><li>bar</li></ul><li>baz</li></ul>" but got "<ol><li>foo</li></ol><ul><li>bar</li></ul>baz"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo]<ul><li>bar</ul>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>[bar]</ul>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ol><li>bar</li></ol><li>baz</li></ul>" but got "<ul><li>foo<ol><li>bar</li></ol>baz</li></ul>"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>[bar]</ul>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>[bar]</ul>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>bar</ul>[baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><ul><li>bar</li></ul></ul><ol><li>baz</li></ol>" but got "foo<ul><li>bar</li></ul><ol><li>baz</li></ol>"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>bar</ul>[baz]</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>foo<ul><li>bar</ul>[baz]</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><ul><li>baz</li></ul>" but got "<ol><li>foo</li><li>bar</li></ol>baz"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul><li>[foo<ul><li>bar]</ul>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz]</ul>quz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz]</ul>quz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz</ul>quz]" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "foo<ol><li>bar</ol><ul><li>[baz</ul>quz]" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "foo<ul><li>[bar]</ul><ol><li>baz</ol>quz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "foo<ul><li>[bar]</ul><ol><li>baz</ol>quz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "[foo<ul><li>bar]</ul><ol><li>baz</ol>quz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "[foo<ul><li>bar]</ul><ol><li>baz</ol>quz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "[foo]<blockquote>bar</blockquote>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "[foo]<blockquote>bar</blockquote>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "foo<blockquote>[bar]</blockquote>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "foo<blockquote>[bar]</blockquote>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "[foo<blockquote>bar]</blockquote>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol>baz" but got "<ol><li>foo</li><li>bar</li></ol>baz"
[FAIL] [["insertorderedlist",""]] "[foo<blockquote>bar]</blockquote>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "[foo<blockquote>bar]</blockquote>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</ol><blockquote>[bar]</blockquote>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol>baz" but got "<ol><li>foo</li><li>bar</li></ol>baz"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</ol><blockquote>[bar]</blockquote>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</ol><blockquote>[bar]</blockquote>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "[foo]<blockquote><ol><li>bar</ol></blockquote>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><blockquote><ol><li>bar</li></ol></blockquote>baz" but got "<blockquote><ol><li>foo</li><li>bar</li></ol></blockquote>baz"
[FAIL] [["insertorderedlist",""]] "[foo]<blockquote><ol><li>bar</ol></blockquote>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "[foo]<blockquote><ol><li>bar</ol></blockquote>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "foo<blockquote>[bar]<br>baz</blockquote>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "foo<blockquote>[bar]<br>baz</blockquote>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "[foo<blockquote>bar]<br>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><blockquote>baz</blockquote>" but got "<ol><li>foo</li><li>bar</li></ol><blockquote>baz</blockquote>"
[FAIL] [["insertorderedlist",""]] "[foo<blockquote>bar]<br>baz</blockquote>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "[foo<blockquote>bar]<br>baz</blockquote>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</ol><blockquote>[bar]<br>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><blockquote>baz</blockquote>" but got "<ol><li>foo</li><li>bar</li></ol><blockquote>baz</blockquote>"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</ol><blockquote>[bar]<br>baz</blockquote>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ol><li>foo</ol><blockquote>[bar]<br>baz</blockquote>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><blockquote><p>bar</p></blockquote><p>baz</p>" but got "<p><ol><li>foo</li></ol></p><blockquote><p>bar</p></blockquote><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><blockquote><p>bar</p></blockquote><p>baz</p>" but got "<p><ol><li>foo</li></ol></p><blockquote><p>bar</p></blockquote><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo]<blockquote><p>bar</blockquote><p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>foo<blockquote><p>[bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><blockquote><ol><li>bar</li></ol></blockquote><p>baz</p>" but got "<p>foo</p><blockquote><p><ol><li>bar</li></ol></p></blockquote><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>foo<blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>foo<blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>foo<blockquote><p>[bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><blockquote><ol><li>bar</li></ol></blockquote><p>baz</p>" but got "<p>foo</p><blockquote><p><ol><li>bar</li></ol></p></blockquote><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>foo<blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>foo<blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo<blockquote><p>bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><p>baz</p>" but got "<p><ol><li>foo</li><li>bar</li></ol></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo<blockquote><p>bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo<blockquote><p>bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo<blockquote><p>bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><p>baz</p>" but got "<p><ol><li>foo</li><li>bar</li></ol></p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo<blockquote><p>bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo<blockquote><p>bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo</ol><blockquote><p>[bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><p>baz</p>" but got "<ol><li>foo</li><li>bar</li></ol><p>baz</p>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo</ol><blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo</ol><blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo</ol><blockquote><p>[bar]</blockquote><p>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><ol><li>bar</li></ol></ol><p>baz</p>" but got "<ol><li>foo</li><li>bar</li></ol><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo</ol><blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo</ol><blockquote><p>[bar]</blockquote><p>baz" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul id=abc><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul id=abc><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["insertorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["insertorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["insertorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["insertorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li></ul><ol><li><font color=\\"#0000ff\\">bar</font></li></ol><ul style=\\"color:rgb(0, 0, 255)\\"><li>baz</li></ul>" but got "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li></ul><ol><li><span style=\\"color:rgb(0, 0, 255)\\">bar</span></li></ol><ul style=\\"color:rgb(0, 0, 255)\\"><li>baz</li></ul>"
[FAIL] [["stylewithcss","false"],["insertorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["insertorderedlist",""]] "<ul style=color:blue><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["insertorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"text-indent:1em\\"><li>foo</li></ul><ol><li>bar</li></ol><ul style=\\"text-indent:1em\\"><li>baz</li></ul>" but got "<ul style=\\"text-indent:1em\\"><li>foo</li></ul><ol><li><span style=\\"text-indent:1em\\">bar</span></li></ol><ul style=\\"text-indent:1em\\"><li>baz</li></ul>"
[FAIL] [["stylewithcss","true"],["insertorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["insertorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["insertorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"text-indent:1em\\"><li>foo</li></ul><ol><li>bar</li></ol><ul style=\\"text-indent:1em\\"><li>baz</li></ul>" but got "<ul style=\\"text-indent:1em\\"><li>foo</li></ul><ol><li><span style=\\"text-indent:1em\\">bar</span></li></ol><ul style=\\"text-indent:1em\\"><li>baz</li></ul>"
[FAIL] [["stylewithcss","false"],["insertorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["insertorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>[bar]<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul id=abc><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul id=abc><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["insertorderedlist",""]] "<ul style=color:blue><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["insertorderedlist",""]] "<ul style=color:blue><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["insertorderedlist",""]] "<ul style=color:blue><li>[foo]<li>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><font color=\\"#0000ff\\">foo</font></li></ol><ul style=\\"color:rgb(0, 0, 255)\\"><li>bar</li><li>baz</li></ul>" but got "<ol><li><span style=\\"color:rgb(0, 0, 255)\\">foo</span></li></ol><ul style=\\"color:rgb(0, 0, 255)\\"><li>bar</li><li>baz</li></ul>"
[FAIL] [["stylewithcss","false"],["insertorderedlist",""]] "<ul style=color:blue><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["insertorderedlist",""]] "<ul style=color:blue><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul style=text-indent:1em><li>[foo]<li>bar<li>baz</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><ul style=\\"text-indent:1em\\"><li>bar</li><li>baz</li></ul>" but got "<ol><li><span style=\\"text-indent:1em\\">foo</span></li></ol><ul style=\\"text-indent:1em\\"><li>bar</li><li>baz</li></ul>"
[FAIL] [["insertorderedlist",""]] "<ul style=text-indent:1em><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul style=text-indent:1em><li>[foo]<li>bar<li>baz</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul id=abc><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul id=abc><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["insertorderedlist",""]] "<ul style=color:blue><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["insertorderedlist",""]] "<ul style=color:blue><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["insertorderedlist",""]] "<ul style=color:blue><li>foo<li>bar<li>[baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li><li>bar</li></ul><ol><li><font color=\\"#0000ff\\">baz</font></li></ol>" but got "<ul style=\\"color:rgb(0, 0, 255)\\"><li>foo</li><li>bar</li></ul><ol><li><span style=\\"color:rgb(0, 0, 255)\\">baz</span></li></ol>"
[FAIL] [["stylewithcss","false"],["insertorderedlist",""]] "<ul style=color:blue><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["insertorderedlist",""]] "<ul style=color:blue><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>bar<li>[baz]</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul style=\\"text-indent:1em\\"><li>foo</li><li>bar</li></ul><ol><li>baz</li></ol>" but got "<ul style=\\"text-indent:1em\\"><li>foo</li><li>bar</li></ul><ol><li><span style=\\"text-indent:1em\\">baz</span></li></ol>"
[FAIL] [["insertorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["insertorderedlist",""]] "<ul style=text-indent:1em><li>foo<li>bar<li>[baz]</ul>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo</ol> <p>[bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li> <li>bar</li></ol>" but got "<ol><li>foo</li><li>bar</li></ol> "
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo</ol> <p>[bar]" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo</ol> <p>[bar]" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo</ol> <p>[bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li> <li>bar</li></ol>" but got "<ol><li>foo</li><li>bar</li></ol> "
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo</ol> <p>[bar]" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo</ol> <p>[bar]" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo]</p> <ol><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li> <li>bar</li></ol>" but got " <ol><li>foo</li><li>bar</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo]</p> <ol><li>bar</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<p>[foo]</p> <ol><li>bar</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo]</p> <ol><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li> <li>bar</li></ol>" but got " <ol><li>foo</li><li>bar</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo]</p> <ol><li>bar</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<p>[foo]</p> <ol><li>bar</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo</ol> <p>[bar]</p> <ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li> <li>bar</li> <li>baz</li></ol>" but got "<ol><li>foo</li><li>bar</li><li>baz</li></ol>  "
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo</ol> <p>[bar]</p> <ol><li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","div"],["insertorderedlist",""]] "<ol><li>foo</ol> <p>[bar]</p> <ol><li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo</ol> <p>[bar]</p> <ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li> <li>bar</li> <li>baz</li></ol>" but got "<ol><li>foo</li><li>bar</li><li>baz</li></ol>  "
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo</ol> <p>[bar]</p> <ol><li>baz</ol>" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["defaultparagraphseparator","p"],["insertorderedlist",""]] "<ol><li>foo</ol> <p>[bar]</p> <ol><li>baz</ol>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["insertorderedlist",""],["inserttext","abc"]] "<div>{}</div>": execCommand("insertorderedlist", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["insertorderedlist",""],["inserttext","abc"]] "<div>{}</div>": execCommand("inserttext", false, "abc") return value
  assert_equals: expected true but got false
[FAIL] [["insertorderedlist",""],["inserttext","abc"]] "<div>{}</div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div></div>" not in array ["<div><ol><li>abc</li></ol></div>", "<div><ol><li>abc<br></li></ol></div>"]
[FAIL] [["insertorderedlist",""],["inserttext","abc"]] "<div>{}</div>" queryCommandState("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertorderedlist",""],["inserttext","abc"]] "<div>{}</div>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "true" but got "false"
[FAIL] [["insertorderedlist",""],["inserttext","abc"]] "<p>{}</p>": execCommand("insertorderedlist", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["insertorderedlist",""],["inserttext","abc"]] "<p>{}</p>": execCommand("inserttext", false, "abc") return value
  assert_equals: expected true but got false
[FAIL] [["insertorderedlist",""],["inserttext","abc"]] "<p>{}</p>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<p></p>" not in array ["<ol><li>abc</li></ol>", "<ol><li>abc<br></li></ol>"]
[FAIL] [["insertorderedlist",""],["inserttext","abc"]] "<p>{}</p>" queryCommandState("insertorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertorderedlist",""],["inserttext","abc"]] "<p>{}</p>" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "true" but got "false"
[FAIL] [["insertorderedlist",""],["inserttext","abc"]] "<p>{}<br></p>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<p><ol><li>abc</li></ol></p>" not in array ["<ol><li>abc</li></ol>", "<ol><li>abc<br></li></ol>"]
[FAIL] [["insertorderedlist",""]] "<div id=\\"a\\">a[bc</div><div id=\\"b\\">de]f</div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div id=\\"a\\"><ol><li>abc</li><li>def</li></ol></div>" not in array ["<div><ol><li id=\\"a\\">abc</li><li id=\\"b\\">def</li></ol></div>", "<ol><li id=\\"a\\">abc</li><li id=\\"b\\">def</li></ol>"]
[FAIL] [["insertorderedlist",""]] "<div class=\\"a\\">a[bc</div><div class=\\"b\\">de]f</div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div class=\\"a\\"><ol><li>abc</li><li>def</li></ol></div>" not in array ["<div><ol><li class=\\"a\\">abc</li><li class=\\"b\\">def</li></ol></div>", "<ol><li class=\\"a\\">abc</li><li class=\\"b\\">def</li></ol>"]
[FAIL] [["insertorderedlist",""]] "<div title=\\"a\\">a[bc</div><div title=\\"b\\">de]f</div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div title=\\"a\\"><ol><li>abc</li><li>def</li></ol></div>" not in array ["<div><ol><li title=\\"a\\">abc</li><li title=\\"b\\">def</li></ol></div>", "<ol><li title=\\"a\\">abc</li><li title=\\"b\\">def</li></ol>"]
Harness: the test ran to completion.
